pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.6.6",
    "reportlab>=4.3.1",
]
